+++ crdesigner/map_conversion/map_conversion_interface.py
@@ -9,6 +9,8 @@ from commonroad.scenario.scenario import Scenario
 from commonroad.common.file_reader import CommonRoadFileReader
 
 from crdesigner.config.opendrive_config import open_drive_config
+from crdesigner.config.lanelet2_config import lanelet2_config
+from crdesigner.config.general_config import general_config
 from crdesigner.map_conversion.opendrive.opendrive_parser.parser import parse_opendrive
 from crdesigner.map_conversion.opendrive.opendrive_conversion.network import Network
 
@@ -31,25 +33,26 @@ from crdesigner.map_conversion.osm2cr.converter_modules.cr_operations.export imp
 from crdesigner.map_conversion.osm2cr.converter_modules.converter import GraphScenario
 
 
-def lanelet_to_commonroad( input_file: str) -> Scenario:
+def lanelet_to_commonroad(input_file: str, general_conf: general_config = general_config,
+                          lanelet2_conf: lanelet2_config = lanelet2_config) -> Scenario:
     """
     Converts lanelet/lanelet2 file to CommonRoad
 
     :param input_file: Path to lanelet/lanelet2 file
-    :param general_config: General config parameters.
-    :param lanelet2_config: Lanelet2 config parameters.
+    :param general_conf: General config parameters.
+    :param lanelet2_conf: Lanelet2 config parameters.
     :return: CommonRoad scenario
     """
-    parser = Lanelet2Parser(etree.parse(input_file).getroot())
+    parser = Lanelet2Parser(etree.parse(input_file).getroot(), lanelet2_conf)
     lanelet2_content = parser.parse()
 
-    lanelet2_converter = Lanelet2CRConverter()
+    lanelet2_converter = Lanelet2CRConverter(lanelet2_conf, general_conf)
     scenario = lanelet2_converter(lanelet2_content)
 
     return scenario
 
 
-def commonroad_to_lanelet(input_file: str, output_name: str):
+def commonroad_to_lanelet(input_file: str, output_name: str, config: lanelet2_config = lanelet2_config):
     """
     Converts CommonRoad map to lanelet format
 
@@ -68,7 +71,7 @@ def commonroad_to_lanelet(input_file: str, output_name: str):
         )
         return
 
-    l2osm = CR2LaneletConverter()
+    l2osm = CR2LaneletConverter(config=config)
     osm = l2osm(scenario)
     with open(f"{output_name}", "wb") as file_out:
         file_out.write(
@@ -78,13 +81,14 @@ def commonroad_to_lanelet(input_file: str, output_name: str):
         )
 
 
-def opendrive_to_commonroad(input_file: str) -> Scenario:
+def opendrive_to_commonroad(input_file: str, general_conf: general_config = general_config,
+                            odr_conf: open_drive_config = open_drive_config) -> Scenario:
     """
     Converts OpenDRIVE file to CommonRoad
 
     :param input_file: Path to OpenDRIVE file
-    :param general_config: General config parameters.
-    :param odr_config: OpenDRIVE config parameters.
+    :param general_conf: General config parameters.
+    :param odr_conf: OpenDRIVE config parameters.
     :return: CommonRoad scenario
     """
     opendrive = parse_opendrive(input_file)
@@ -93,7 +97,7 @@ def opendrive_to_commonroad(input_file: str) -> Scenario:
     for index in range(len(road_network._traffic_lights)):
         road_network._traffic_lights[index]._traffic_light_id = \
             abs(road_network._traffic_lights[index].traffic_light_id)
-    return road_network.export_commonroad_scenario()
+    return road_network.export_commonroad_scenario(general_conf, odr_conf)
 
 
 def sumo_to_commonroad(input_file: str) -> Scenario: