+++ crdesigner/map_conversion/sumo_map/cr2sumo/traffic_light.py
@@ -22,6 +22,7 @@ class TrafficLightEncoder:
         program_id = f"tl_program_{self.index}"
         traffic_light_id = str(node.id)
         node.type = NodeType.TRAFFIC_LIGHT
+
         time_offset = max(tl.traffic_light_cycle.time_offset for tl in traffic_lights)
         tls_program = TLSProgram(traffic_light_id, time_offset * self.conf.dt, program_id)
 
@@ -86,8 +87,7 @@ def _sync_traffic_light_cycles(traffic_lights: List[TrafficLight]) -> List[List[
     :param traffic_lights:
     :return: list of synchronized states
     """
-    time_steps = np.lcm.reduce([sum(cycle.duration for cycle in tl.traffic_light_cycle.cycle_elements)
-                                for tl in traffic_lights])
+    time_steps = np.lcm.reduce([sum(cycle.duration for cycle in tl.traffic_light_cycle.cycle_elements) for tl in traffic_lights])
     states = np.array([_cycles_to_states(traffic_light.traffic_light_cycle.cycle_elements, time_steps)
                        for traffic_light in traffic_lights]).T
     res: List[List[TrafficLightCycleElement]] = []