+++ crdesigner/ui/gui/controller/toolboxes/converter_toolbox/map_conversion_controller.py
@@ -2,6 +2,7 @@ import warnings
 from lxml import etree
 from typing import Callable, Optional
 
+
 from crdesigner.map_conversion.lanelet2.cr2lanelet import CR2LaneletConverter
 from crdesigner.map_conversion.lanelet2.lanelet2_parser import Lanelet2Parser
 from crdesigner.map_conversion.lanelet2.lanelet2cr import Lanelet2CRConverter
@@ -36,7 +37,7 @@ class RequestRunnable(QRunnable):
 
     def run(self):
         self.fun()
-        QMetaObject.invokeMethod(self.mapConversionToolboxController, "stopSpinner", Qt.QueuedConnection,
+        QMetaObject.invokeMethod(self.mapConversionToolboxController, "stop_spinner", Qt.QueuedConnection,
                                  Q_ARG(str, "Conversion Ended"))
 
 
@@ -111,7 +112,6 @@ class MapConversionToolboxController(QDockWidget):
         filename = select_local_file(self, "OSM", "osm")
         if filename != "":
             self.osm_file = filename
-
 
     def hidden_osm_conversion(self, graph: rg.Graph) -> None:
         """
@@ -207,7 +207,6 @@ class MapConversionToolboxController(QDockWidget):
         print(data)
         self.scenario_model.stop_spinner()
         self.converter_toolbox.Spinner.stop()
-
 
     def start_spinner(self, spinner: QtWaitingSpinner):
         if spinner.is_spinning():